requirements.txt
flask>=3.0,<4.0
prometheus-client>=0.21,<1.0
requests>=2.31,<3.0
gunicorn>=22.0,<23.0
python-dotenv>=1.0,<2.0
